feat: add search from bytes option and refactor
- search from bytes lets us do fetch API instead of needing to embed the serialized binary with include bytes
- refactored binary mian.rs and library lib.rs to make  more sense
